Rename templates actions file
