+++ src/pages/Experiments/Experiment/ExperimentHeader/NewTemplateModal/NewTemplateModalContainer.js
@@ -7,7 +7,7 @@ import * as TEMPLATES_TYPES from 'store/templates/actionTypes';
 import { useParams } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { createTemplateRequest } from 'store/templates/actions';
+import { createTemplateRequest } from 'store/templates/templates.actions';
 import { hideNewTemplateModal } from 'store/ui/actions';
 
 import { useIsLoading } from 'hooks';
